[Feature #160478873] Get all orders
